Refactor shadowfork xtask helpers
